requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
requests
pyyaml
jsoncomparedeep
jsonpath